Added ability to forward query strings to tests
This is an option on the run window.

Also cleaned up the option handling and added URL handling for all
of the options on the run page.  Added documentation for the URL
parameters to the instructions page.

All changes are backward compatible (in that I did not change any of the
existing names / ids / classes on the controls) just in case someone was
relying upon those.