Added code to support showing subtest output
As per an IRC discussion that this could be useful for some situations.

This PR also adds support for markdown in subtest names and messages.
In that way tests and test reports can be more generally useful with
links to specs (for example).  The annotation-model tests in WPT take
advantage of this feature now.